Create a points field in portfolio and add a system to keep it updated

+++ market-league-back-end/internal/models/leaderboard_entry.go
@@ -2,6 +2,6 @@ package models
 
 // LeaderboardEntry represents an entry in the league leaderboard.
 type LeaderboardEntry struct {
-	Username   string  `json:"username"`
-	TotalValue float64 `json:"total_value"`
+	Username   string `json:"username"`
+	TotalValue int    `json:"total_value"`
 }
